[#159] Renaming almost done
