+++ datastock/__init__.py
@@ -4,7 +4,7 @@ from .version import __version__
 
 from . import _generic_check
 from ._generic_utils_plot import *
-from ._class import DataStock
+from ._class04 import Plots as Collection
 from ._saveload import load, get_files
 from ._direct_calls import *
 from . import tests
