--- datastock/_class.py
@@ -1,3 +1,0 @@
-
-
-from ._class3 import DataStock3 as DataStock
